Time: 126 ms (20.09%), Space: 67.9 MB (36.88%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.java
@@ -19,7 +19,7 @@ class MedianFinder {
             return 0;
         }
         if (small.size() == large.size()) {
-            return (double)(small.peek() + large.peek()) / 2.0;
+            return (small.peek() + large.peek()) / 2.0;
         } else {
             return small.peek();
         }
